pyproject.toml
[project]
name = "math-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.2",
    "logfire>=4.3.6",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]
